Refactor and Tests.
Refactored a lot of the private class variables to not use static.
Was causing a lot of problems when running multiple tests, although
the game will never create more than one logic, and map class.

Fixed Map class generate tank position to return the actual List of
positions and not just the current tank position.

Added tests to test all major functions in every class.

Refactored a lot of the code and wrote new tests.
A lot of the static fields were causing problems when running multiple tests.